\name{fixed.effects}
\alias{fixed.effects}
\alias{fixef}
\docType{import}
\title{Import from package \pkg{nlme}}
\description{
  The \code{fixed.effects} and \code{fixef} methods are
  imported from package \pkg{nlme}.
Help is available here:  \code{\link[nlme:fixed.effects]{nlme::fixed.effects}}.
}
