\name{glance.coxme}
\Rdversion{1.1}
\alias{glance.coxme}
\title{
Glance at a coxme object
}
\description{
\code{glance()} returns a one-row summary of a model.
}
\usage{
\method{glance}{coxme}(x, ...)
}
\arguments{
  \item{x}{
    an object of class \code{coxme}, from the fit of a mixed effects
    survival model.
  }
  \item{\dots}{
    optional arguments passed to \code{\link{generics::glance}}}.
}
\author{
Jason Cory Brunson
}
\seealso{\code{\link{coxme}}}
\keyword{survival}
