\name{print.lmekin}
\alias{print.lmekin}
\title{Print function for lmekin}
\description{Print out the result of an lmekin fit.}
\usage{
\method{print}{lmekin}(x, ...)
}
\arguments{
  \item{x}{an object of class \code{lmekin}.}
  \item{\dots}{generic arguments to print, unused.}
}
\details{
 The print function current has no options.  This should one day improve.
}
\author{Terry Therneau}
\seealso{\code{\link{lmekin}}}
\keyword{ models }
